@@ -4,12 +4,11 @@
   "private": true,
   "dependencies": {
     "@babel/plugin-transform-react-jsx": "^7.16.7",
-    "@emotion/react": "^11.9.3",
-    "@emotion/styled": "^11.9.3",
-    "@mui/icons-material": "^5.2.4",
+    "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
+    "@mui/icons-material": "^5.14.7",
     "@mui/lab": "^5.0.0-alpha.60",
-    "@mui/material": "^5.9.2",
-    "@mui/styles": "^5.2.3",
+    "@mui/material": "^5.14.7",
     "@mui/x-date-pickers": "^5.0.0-beta.2",
     "@testing-library/jest-dom": "^5.15.0",
     "@testing-library/react": "^11.2.7",
@@ -22,11 +21,11 @@
     "firebase-tools": "^11.30.0",
     "glob-parent": "^6.0.2",
     "jose": "^4.9.2",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-copy-to-clipboard": "^5.0.4",
     "react-dnd": "^16.0.1",
     "react-dnd-html5-backend": "^16.0.1",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-icons": "^4.3.1",
     "react-router-dom": "^6.0.2",
     "react-uuid": "^2.0.0",
