@@ -13,20 +13,7 @@ import SettingsIcon from "@mui/icons-material/Settings";
 import { WatchListRow } from "./ItemRow";
 import { useContext, useState } from "react";
 import { UserWatchlistContext } from "../../../../Context/AuthContext";
-import { makeStyles } from "@mui/styles";
 import { useFirebase } from "../../../../Hooks/useFirebase";
-
-const useStyles = makeStyles((theme) => ({
-  Accordion: {
-    "& .MuiAccordionSummary-root:hover": {
-      cursor: "default",
-    },
-  },
-  Header: {
-    color:
-      theme.palette.type === "dark" ? "secondary" : theme.palette.primary.main,
-  },
-}));
 
 export function WatchlistGroup({
   group,
@@ -39,17 +26,20 @@ export function WatchlistGroup({
     useContext(UserWatchlistContext);
   const { uploadUserWatchlist } = useFirebase();
   const [expandGroup, updateExpandGroup] = useState(group.expanded);
-  const classes = useStyles();
 
   return (
     <Grid item xs={12}>
       <Accordion
-        className={classes.Accordion}
         expanded={expandGroup}
-        square={true}
+        square
         spacing={1}
         id={group.id}
-        disableGutters={true}
+        disableGutters
+        sx={{
+          "& .MuiAccordionSummary-root:hover": {
+            cursor: "default",
+          },
+        }}
       >
         <AccordionSummary
           expandIcon={
@@ -76,7 +66,15 @@ export function WatchlistGroup({
             <Box
               sx={{ display: "flex", flex: "1 1 95%", flexDirection: "row" }}
             >
-              <Typography variant="h6" className={classes.Header}>
+              <Typography
+                variant="h6"
+                sx={{
+                  color: (theme) =>
+                    theme.palette.type === "dark"
+                      ? "secondary"
+                      : theme.palette.primary.main,
+                }}
+              >
                 {group.name}
               </Typography>
             </Box>
