@@ -1,24 +1,13 @@
 import { useContext, useMemo } from "react";
 import { Grid, Paper, Typography, Checkbox } from "@mui/material";
 import { UsersContext } from "../../../Context/AuthContext";
-import { makeStyles } from "@mui/styles";
 import { useFirebase } from "../../../Hooks/useFirebase";
 import { UserLoginUIContext } from "../../../Context/LayoutContext";
-
-const useStyles = makeStyles((theme) => ({
-  Checkbox: {
-    color:
-      theme.palette.type === "dark"
-        ? theme.palette.primary.main
-        : theme.palette.secondary.main,
-  },
-}));
 
 export function TutorialDashboard() {
   const { users, updateUsers } = useContext(UsersContext);
   const { userDataFetch } = useContext(UserLoginUIContext);
   const { updateMainUserDoc } = useFirebase();
-  const classes = useStyles();
   const parentUser = useMemo(() => users.find((i) => i.ParentUser), [users]);
 
   if (!parentUser.settings.layout.hideTutorials && !userDataFetch) {
@@ -56,7 +45,12 @@ export function TutorialDashboard() {
               <Grid item xs={2} align="right">
                 <Typography variant="caption">Hide Help Options</Typography>
                 <Checkbox
-                  className={classes.Checkbox}
+                  sx={{
+                    color: (theme) =>
+                      theme.palette.type === "dark"
+                        ? theme.palette.primary.main
+                        : theme.palette.secondary.main,
+                  }}
                   size="small"
                   onClick={() => {
                     let newUsers = JSON.parse(JSON.stringify(users));
