@@ -15,20 +15,10 @@ import { functions } from "../../firebase";
 import { UsersContext } from "../../Context/AuthContext";
 import { SnackBarDataContext } from "../../Context/LayoutContext";
 import { httpsCallable } from "firebase/functions";
-import { makeStyles } from "@mui/styles";
 import {
   CorpEsiDataContext,
   PersonalESIDataContext,
 } from "../../Context/EveDataContext";
-
-const useStyles = makeStyles((theme) => ({
-  Checkbox: {
-    color:
-      theme.palette.type === "dark"
-        ? theme.palette.primary.main
-        : theme.palette.secondary.main,
-  },
-}));
 
 export function FeedbackIcon() {
   const { users } = useContext(UsersContext);
@@ -47,7 +37,6 @@ export function FeedbackIcon() {
   const [open, setOpen] = useState(false);
   const [inputText, updateInputText] = useState("");
   const [dataDump, updateDataDump] = useState(false);
-  const classes = useStyles();
 
   const parentUser = useMemo(() => {
     return users.find((i) => i.ParentUser);
@@ -79,7 +68,10 @@ export function FeedbackIcon() {
       };
     };
 
-    const call = httpsCallable(functions, "submitUserFeedback-submitUserFeedback");
+    const call = httpsCallable(
+      functions,
+      "submitUserFeedback-submitUserFeedback"
+    );
     call({
       accountID: parentUser.accountID || null,
       response: inputText,
@@ -145,12 +137,17 @@ export function FeedbackIcon() {
           <FormControlLabel
             control={
               <Checkbox
-                className={classes.Checkbox}
                 checked={dataDump}
                 onChange={() => {
                   updateDataDump((prev) => !prev);
                 }}
-                sx={{ marginRight: { xs: "20px", sm: "60px" } }}
+                sx={{
+                  marginRight: { xs: "20px", sm: "60px" },
+                  color: (theme) =>
+                    theme.palette.type === "dark"
+                      ? theme.palette.primary.main
+                      : theme.palette.secondary.main,
+                }}
               />
             }
             label="Include ESI data?"
