@@ -4,26 +4,11 @@ import { ActiveJobContext } from "../../../../../Context/JobContext";
 import { SnackBarDataContext } from "../../../../../Context/LayoutContext";
 import ClearIcon from "@mui/icons-material/Clear";
 import { MdAdd } from "react-icons/md";
-import { makeStyles } from "@mui/styles";
-
-const useStyles = makeStyles((theme) => ({
-  TextField: {
-    "& .MuiFormHelperText-root": {
-      color: theme.palette.secondary.main,
-    },
-  },
-  Autocomplete: {
-    "& .MuiFormHelperText-root": {
-      color: theme.palette.secondary.main,
-    },
-  },
-}));
 
 export function ExtrasList({ setJobModified }) {
   const { activeJob, updateActiveJob } = useContext(ActiveJobContext);
   const [extras, updateExtras] = useState({ text: "", value: 0 });
   const { setSnackbarData } = useContext(SnackBarDataContext);
-  const classes = useStyles();
 
   function handleAdd(event) {
     event.preventDefault();
@@ -147,7 +132,11 @@ export function ExtrasList({ setJobModified }) {
           <Grid container direction="row" item xs={12} spacing={3}>
             <Grid item xs={5}>
               <TextField
-                className={classes.TextField}
+                sx={{
+                  "& .MuiFormHelperText-root": {
+                    color: (theme) => theme.palette.secondary.main,
+                  },
+                }}
                 defaultValue={extras.text}
                 variant="standard"
                 required={true}
@@ -165,7 +154,11 @@ export function ExtrasList({ setJobModified }) {
             </Grid>
             <Grid item xs={5}>
               <TextField
-                className={classes.TextField}
+                sx={{
+                  "& .MuiFormHelperText-root": {
+                    color: (theme) => theme.palette.secondary.main,
+                  },
+                }}
                 defaultValue={extras.value}
                 variant="standard"
                 required={true}
