@@ -15,28 +15,10 @@ import {
 import { useContext, useEffect, useState } from "react";
 import { UsersContext } from "../../../../Context/AuthContext";
 import { listingType } from "../../../../Context/defaultValues";
-import { makeStyles } from "@mui/styles";
 import { useCharAssets } from "../../../../Hooks/useCharAssets";
 import { EveIDsContext } from "../../../../Context/EveDataContext";
 import { useFirebase } from "../../../../Hooks/useFirebase";
 import GLOBAL_CONFIG from "../../../../global-config-app";
-
-const useStyles = makeStyles((theme) => ({
-  TextField: {
-    "& .MuiFormHelperText-root": {
-      color: theme.palette.secondary.main,
-    },
-    "& input::-webkit-clear-button, & input::-webkit-outer-spin-button, & input::-webkit-inner-spin-button":
-      {
-        display: "none",
-      },
-  },
-  Select: {
-    "& .MuiFormHelperText-root": {
-      color: theme.palette.secondary.main,
-    },
-  },
-}));
 
 export function CompactEditJobSettings({ parentUserIndex }) {
   const { users, updateUsers } = useContext(UsersContext);
@@ -55,8 +37,6 @@ export function CompactEditJobSettings({ parentUserIndex }) {
   );
   const [assetLocationEntries, updateAssetLocationEntries] = useState([]);
   const { MARKET_OPTIONS } = GLOBAL_CONFIG;
-
-  const classes = useStyles();
 
   useEffect(() => {
     async function getAsset() {
@@ -79,7 +59,18 @@ export function CompactEditJobSettings({ parentUserIndex }) {
         </Grid>
         <Grid container item xs={12}>
           <Grid item xs={6} sm={4} lg={3}>
-            <FormControl className={classes.Select} fullWidth={true}>
+            <FormControl
+              sx={{
+                "& .MuiFormHelperText-root": {
+                  color: (theme) => theme.palette.secondary.main,
+                },
+                "& input::-webkit-clear-button, & input::-webkit-outer-spin-button, & input::-webkit-inner-spin-button":
+                  {
+                    display: "none",
+                  },
+              }}
+              fullWidth={true}
+            >
               <Select
                 value={marketSelect}
                 variant="standard"
@@ -111,7 +102,18 @@ export function CompactEditJobSettings({ parentUserIndex }) {
             </FormControl>
           </Grid>
           <Grid item xs={6} sm={4} lg={3}>
-            <FormControl className={classes.Select} fullWidth={true}>
+            <FormControl
+              sx={{
+                "& .MuiFormHelperText-root": {
+                  color: (theme) => theme.palette.secondary.main,
+                },
+                "& input::-webkit-clear-button, & input::-webkit-outer-spin-button, & input::-webkit-inner-spin-button":
+                  {
+                    display: "none",
+                  },
+              }}
+              fullWidth={true}
+            >
               <Select
                 value={listingSelect}
                 variant="standard"
@@ -190,7 +192,18 @@ export function CompactEditJobSettings({ parentUserIndex }) {
             {dataLoading ? (
               <CircularProgress color="primary" size="20px" />
             ) : (
-              <FormControl className={classes.Select} fullWidth>
+              <FormControl
+                sx={{
+                  "& .MuiFormHelperText-root": {
+                    color: (theme) => theme.palette.secondary.main,
+                  },
+                  "& input::-webkit-clear-button, & input::-webkit-outer-spin-button, & input::-webkit-inner-spin-button":
+                    {
+                      display: "none",
+                    },
+                }}
+                fullWidth
+              >
                 <Select
                   value={assetLocationSelect}
                   variant="standard"
@@ -240,7 +253,15 @@ export function CompactEditJobSettings({ parentUserIndex }) {
               }
               size="small"
               variant="standard"
-              className={classes.TextField}
+              sx={{
+                "& .MuiFormHelperText-root": {
+                  color: (theme) => theme.palette.secondary.main,
+                },
+                "& input::-webkit-clear-button, & input::-webkit-outer-spin-button, & input::-webkit-inner-spin-button":
+                  {
+                    display: "none",
+                  },
+              }}
               helperText="Citadel Brokers Fee Percentage"
               type="number"
               onBlur={(e) => {
