@@ -18,7 +18,6 @@ import {
 } from "@mui/material";
 import { structureOptions } from "../../../../Context/defaultValues";
 import AddIcon from "@mui/icons-material/Add";
-import { makeStyles } from "@mui/styles";
 import { useState } from "react";
 import { useContext } from "react";
 import { UsersContext } from "../../../../Context/AuthContext";
@@ -28,26 +27,6 @@ import { getAnalytics, logEvent } from "firebase/analytics";
 import { SnackBarDataContext } from "../../../../Context/LayoutContext";
 import systemIDS from "../../../../RawData/systems.json";
 import uuid from "react-uuid";
-
-const useStyles = makeStyles((theme) => ({
-  TextField: {
-    "& .MuiFormHelperText-root": {
-      color: theme.palette.secondary.main,
-    },
-    "& input::-webkit-clear-button, & input::-webkit-outer-spin-button, & input::-webkit-inner-spin-button":
-      {
-        display: "none",
-      },
-  },
-  Autocomplete: {
-    "& .MuiInputBase-input.MuiAutocomplete-input.MuiAutocomplete-inputRoot": {
-      color:
-        theme.palette.type === "dark" ? "black" : theme.palette.secondary.main,
-      borderColor:
-        theme.palette.type === "dark" ? "black" : theme.palette.secondary.main,
-    },
-  },
-}));
 
 export function CompactReactionStrutures({ parentUserIndex }) {
   const { users, updateUsers } = useContext(UsersContext);
@@ -65,7 +44,6 @@ export function CompactReactionStrutures({ parentUserIndex }) {
   );
   const [taxValue, updateTaxValue] = useState("");
   const [systemIDValue, updateSystemIDValue] = useState("");
-  const classes = useStyles();
   const analytics = getAnalytics();
 
   const handleSubmit = (event) => {
@@ -128,7 +106,15 @@ export function CompactReactionStrutures({ parentUserIndex }) {
                         required={true}
                         size="small"
                         variant="standard"
-                        className={classes.TextField}
+                        sx={{
+                          "& .MuiFormHelperText-root": {
+                            color: (theme) => theme.palette.secondary.main,
+                          },
+                          "& input::-webkit-clear-button, & input::-webkit-outer-spin-button, & input::-webkit-inner-spin-button":
+                            {
+                              display: "none",
+                            },
+                        }}
                         helperText="Name"
                         type="text"
                         onBlur={(e) => {
@@ -142,7 +128,15 @@ export function CompactReactionStrutures({ parentUserIndex }) {
                     </Grid>
                     <Grid item xs={6} sx={{ paddingLeft: "5px" }}>
                       <FormControl
-                        className={classes.TextField}
+                        sx={{
+                          "& .MuiFormHelperText-root": {
+                            color: (theme) => theme.palette.secondary.main,
+                          },
+                          "& input::-webkit-clear-button, & input::-webkit-outer-spin-button, & input::-webkit-inner-spin-button":
+                            {
+                              display: "none",
+                            },
+                        }}
                         fullWidth={true}
                       >
                         <Select
@@ -170,7 +164,15 @@ export function CompactReactionStrutures({ parentUserIndex }) {
                     </Grid>
                     <Grid item xs={6} sx={{ paddingRight: "5px" }}>
                       <FormControl
-                        className={classes.TextField}
+                        sx={{
+                          "& .MuiFormHelperText-root": {
+                            color: (theme) => theme.palette.secondary.main,
+                          },
+                          "& input::-webkit-clear-button, & input::-webkit-outer-spin-button, & input::-webkit-inner-spin-button":
+                            {
+                              display: "none",
+                            },
+                        }}
                         fullWidth={true}
                       >
                         <Select
@@ -198,7 +200,15 @@ export function CompactReactionStrutures({ parentUserIndex }) {
                     </Grid>
                     <Grid item xs={6} sx={{ paddingLeft: "5px" }}>
                       <FormControl
-                        className={classes.TextField}
+                        sx={{
+                          "& .MuiFormHelperText-root": {
+                            color: (theme) => theme.palette.secondary.main,
+                          },
+                          "& input::-webkit-clear-button, & input::-webkit-outer-spin-button, & input::-webkit-inner-spin-button":
+                            {
+                              display: "none",
+                            },
+                        }}
                         fullWidth={true}
                       >
                         <Select
@@ -226,14 +236,21 @@ export function CompactReactionStrutures({ parentUserIndex }) {
                     </Grid>
                     <Grid item xs={6} sx={{ paddingRight: "5px" }}>
                       <FormControl
-                        className={classes.TextField}
+                        sx={{
+                          "& .MuiFormHelperText-root": {
+                            color: (theme) => theme.palette.secondary.main,
+                          },
+                          "& input::-webkit-clear-button, & input::-webkit-outer-spin-button, & input::-webkit-inner-spin-button":
+                            {
+                              display: "none",
+                            },
+                        }}
                         fullWidth={true}
                       >
                         <TextField
                           required={true}
                           size="small"
                           variant="standard"
-                          className={classes.TextField}
                           helperText="Installation Tax %"
                           inputProps={{
                             step: "0.01",
@@ -247,7 +264,15 @@ export function CompactReactionStrutures({ parentUserIndex }) {
                     </Grid>
                     <Grid item xs={6} sx={{ paddingLeft: "5px" }}>
                       <FormControl
-                        className={classes.TextField}
+                        sx={{
+                          "& .MuiFormHelperText-root": {
+                            color: (theme) => theme.palette.secondary.main,
+                          },
+                          "& input::-webkit-clear-button, & input::-webkit-outer-spin-button, & input::-webkit-inner-spin-button":
+                            {
+                              display: "none",
+                            },
+                        }}
                         fullWidth={true}
                       >
                         <Autocomplete
@@ -267,7 +292,16 @@ export function CompactReactionStrutures({ parentUserIndex }) {
                             <TextField
                               {...params}
                               size="small"
-                              className={classes.Autocomplete}
+                              sx={{
+                                color: (theme) =>
+                                  theme.palette.type === "dark"
+                                    ? "black"
+                                    : theme.palette.secondary.main,
+                                borderColor: (theme) =>
+                                  theme.palette.type === "dark"
+                                    ? "black"
+                                    : theme.palette.secondary.main,
+                              }}
                               margin="none"
                               variant="standard"
                               style={{ borderRadius: "5px" }}
@@ -362,10 +396,10 @@ export function CompactReactionStrutures({ parentUserIndex }) {
                             <Typography variant="body1">{rigText}</Typography>
                           </Grid>
                           <Grid item xs={6}>
-                            <Typography>{`${entry.tax||0}%`}</Typography>
+                            <Typography>{`${entry.tax || 0}%`}</Typography>
                           </Grid>
                           <Grid item xs={6}>
-                            <Typography>{systemName}</Typography> 
+                            <Typography>{systemName}</Typography>
                           </Grid>
                         </Grid>
                       </CardContent>{" "}
