@@ -1,1246 +1,6 @@
-export let jobArrayDefault = [
-  {
-    groupID: null,
-    buildVer: "0.6.0",
-    jobType: 1,
-    name: "Nova Light Missile",
-    jobID: "job-1642030534709",
-    isSnapshot: false,
-    jobStatus: 2,
-    volume: 0.015,
-    itemID: 213,
-    maxProductionLimit: 600,
-    runCount: 50,
-    jobCount: 1,
-    bpME: 0,
-    bpTE: 0,
-    rigType: 0,
-    systemType: 1,
-    apiJobs: new Set(),
-    apiOrders: new Set(),
-    apiTransactions: new Set(),
-    parentJob: [],
-    build: {
-      products: {
-        totalQuantity: 5000,
-        quantityPerJob: 5000,
-      },
-      costs: {
-        totalPurchaseCost: 41300,
-        extrasCosts: [],
-        extrasTotal: 0,
-        linkedJobs: [],
-        installCosts: 0,
-      },
-      sale: {
-        totalSold: 0,
-        totalSale: 0,
-        markUp: 0,
-        marketOrders: [],
-        transactions: [],
-        brokersFee: [],
-      },
-      materials: [
-        {
-          name: "Tritanium",
-          quantity: 2700,
-          jobType: 0,
-          typeID: 34,
-          volume: 0.01,
-          childJob: [],
-          purchasing: [
-            {
-              id: 1642986913280,
-              itemCount: 2700,
-              itemCost: 4,
-            },
-          ],
-          quantityPurchased: 2700,
-          purchasedCost: 10800,
-          purchaseComplete: true,
-        },
-        {
-          volume: 0.01,
-          jobType: 0,
-          quantity: 800,
-          typeID: 35,
-          name: "Pyerite",
-          childJob: [],
-          purchasing: [{ id: 1642986918001, itemCost: 13, itemCount: 800 }],
-          quantityPurchased: 800,
-          purchasedCost: 10400,
-          purchaseComplete: true,
-        },
-        {
-          jobType: 0,
-          name: "Mexallon",
-          typeID: 36,
-          volume: 0.01,
-          quantity: 300,
-          childJob: [],
-          purchasing: [{ id: 1642986931990, itemCost: 67, itemCount: 300 }],
-          quantityPurchased: 300,
-          purchasedCost: 20100,
-          purchaseComplete: true,
-        },
-      ],
-      buildChar: "ABC123",
-      time: 10200,
-    },
-    rawData: {
-      materials: [
-        {
-          name: "Tritanium",
-          quantity: 54,
-          jobType: 0,
-          typeID: 34,
-          volume: 0.01,
-        },
-        {
-          volume: 0.01,
-          jobType: 0,
-          quantity: 16,
-          typeID: 35,
-          name: "Pyerite",
-        },
-        {
-          jobType: 0,
-          name: "Mexallon",
-          typeID: 36,
-          volume: 0.01,
-          quantity: 6,
-        },
-      ],
-      products: [
-        {
-          typeID: 213,
-          quantity: 100,
-        },
-      ],
-      time: 300,
-    },
-    structureType: 0,
-    skills: [
-      {
-        level: 1,
-        typeID: 3380,
-      },
-    ],
-    layout: {
-      localMarketDisplay: null,
-      localOrderDisplay: null,
-      esiJobTab: null,
-    },
-  },
-  {
-    groupID: null,
-    buildVer: "0.2.0",
-    jobType: 2,
-    name: "Scandium Metallofullerene",
-    jobID: "job-1642710994021",
-    jobStatus: 1,
-    volume: 0.65,
-    itemID: 30308,
-    maxProductionLimit: 1000,
-    runCount: 12,
-    jobCount: 1,
-    bpME: 0,
-    bpTE: 0,
-    rigType: 0,
-    systemType: 1,
-    apiJobs: new Set(),
-    apiOrders: new Set(),
-    apiTransactions: new Set(),
-    parentJob: ["job-1642710724535"],
-    build: {
-      products: {
-        totalQuantity: 1920,
-        quantityPerJob: 1920,
-      },
-      costs: {
-        totalPurchaseCost: 0,
-        extrasCosts: [],
-        extrasTotal: 0,
-        linkedJobs: [],
-        installCosts: 0,
-      },
-      sale: {
-        totalSold: 0,
-        totalSale: 0,
-        markUp: 0,
-        marketOrders: [],
-        transactions: [],
-        brokersFee: [],
-      },
-      materials: [
-        {
-          quantity: 60,
-          jobType: 1,
-          typeID: 4247,
-          name: "Helium Fuel Block",
-          volume: 5,
-          childJob: [],
-          purchasing: [],
-          quantityPurchased: 0,
-          purchasedCost: 0,
-          purchaseComplete: false,
-        },
-        {
-          jobType: 0,
-          typeID: 39,
-          quantity: 300,
-          volume: 0.01,
-          name: "Zydrine",
-          childJob: [],
-          purchasing: [],
-          quantityPurchased: 0,
-          purchasedCost: 0,
-          purchaseComplete: false,
-        },
-        {
-          volume: 2,
-          typeID: 30373,
-          name: "Fullerite-C72",
-          quantity: 1200,
-          jobType: 0,
-          childJob: [],
-          purchasing: [],
-          quantityPurchased: 0,
-          purchasedCost: 0,
-          purchaseComplete: false,
-        },
-        {
-          typeID: 30375,
-          name: "Fullerite-C28",
-          quantity: 1200,
-          jobType: 0,
-          volume: 2,
-          childJob: [],
-          purchasing: [],
-          quantityPurchased: 0,
-          purchasedCost: 0,
-          purchaseComplete: false,
-        },
-      ],
-      buildChar: "ABC123",
-      time: 103680,
-    },
-    rawData: {
-      materials: [
-        {
-          quantity: 5,
-          jobType: 1,
-          typeID: 4247,
-          name: "Helium Fuel Block",
-          volume: 5,
-        },
-        {
-          jobType: 0,
-          typeID: 39,
-          quantity: 25,
-          volume: 0.01,
-          name: "Zydrine",
-        },
-        {
-          volume: 2,
-          typeID: 30373,
-          name: "Fullerite-C72",
-          quantity: 100,
-          jobType: 0,
-        },
-        {
-          typeID: 30375,
-          name: "Fullerite-C28",
-          quantity: 100,
-          jobType: 0,
-          volume: 2,
-        },
-      ],
-      products: [
-        {
-          typeID: 30308,
-          quantity: 160,
-        },
-      ],
-      time: 10800,
-    },
-    structureType: 1,
-    skills: [
-      {
-        typeID: 45746,
-        level: 3,
-      },
-    ],
-    layout: {
-      localMarketDisplay: null,
-      localOrderDisplay: null,
-      esiJobTab: null,
-    },
-  },
-  {
-    groupID: null,
-    jobType: 2,
-    name: "Graphene Nanoribbons",
-    jobID: "job-1642711045845",
-    jobStatus: 1,
-    volume: 1.5,
-    itemID: 30309,
-    maxProductionLimit: 1000,
-    runCount: 5,
-    jobCount: 1,
-    bpME: 0,
-    bpTE: 0,
-    rigType: 0,
-    systemType: 1,
-    apiJobs: new Set(),
-    apiOrders: new Set(),
-    apiTransactions: new Set(),
-    parentJob: ["job-1642710724535"],
-    build: {
-      products: {
-        totalQuantity: 600,
-        quantityPerJob: 600,
-      },
-      costs: {
-        totalPurchaseCost: 0,
-        extrasCosts: [],
-        extrasTotal: 0,
-        linkedJobs: [],
-        installCosts: 0,
-      },
-      sale: {
-        totalSold: 0,
-        totalSale: 0,
-        markUp: 0,
-        marketOrders: [],
-        transactions: [],
-        brokersFee: [],
-      },
-      materials: [
-        {
-          quantity: 25,
-          name: "Nitrogen Fuel Block",
-          typeID: 4051,
-          jobType: 1,
-          volume: 5,
-          childJob: [],
-          purchasing: [],
-          quantityPurchased: 0,
-          purchasedCost: 0,
-          purchaseComplete: false,
-        },
-        {
-          name: "Fullerite-C28",
-          volume: 2,
-          typeID: 30375,
-          quantity: 500,
-          jobType: 0,
-          childJob: [],
-          purchasing: [],
-          quantityPurchased: 0,
-          purchasedCost: 0,
-          purchaseComplete: false,
-        },
-        {
-          quantity: 500,
-          volume: 5,
-          typeID: 30376,
-          name: "Fullerite-C32",
-          jobType: 0,
-          childJob: [],
-          purchasing: [],
-          quantityPurchased: 0,
-          purchasedCost: 0,
-          purchaseComplete: false,
-        },
-        {
-          typeID: 38,
-          volume: 0.01,
-          quantity: 2000,
-          name: "Nocxium",
-          jobType: 0,
-          childJob: [],
-          purchasing: [],
-          quantityPurchased: 0,
-          purchasedCost: 0,
-          purchaseComplete: false,
-        },
-      ],
-      buildChar: "ABC123",
-      time: 43200,
-    },
-    rawData: {
-      materials: [
-        {
-          quantity: 5,
-          name: "Nitrogen Fuel Block",
-          typeID: 4051,
-          jobType: 1,
-          volume: 5,
-        },
-        {
-          name: "Fullerite-C28",
-          volume: 2,
-          typeID: 30375,
-          quantity: 100,
-          jobType: 0,
-        },
-        {
-          quantity: 100,
-          volume: 5,
-          typeID: 30376,
-          name: "Fullerite-C32",
-          jobType: 0,
-        },
-        {
-          typeID: 38,
-          volume: 0.01,
-          quantity: 400,
-          name: "Nocxium",
-          jobType: 0,
-        },
-      ],
-      products: [
-        {
-          typeID: 30309,
-          quantity: 120,
-        },
-      ],
-      time: 10800,
-    },
-    structureType: 1,
-    skills: [
-      {
-        level: 3,
-        typeID: 45746,
-      },
-    ],
-    layout: {
-      localMarketDisplay: null,
-      localOrderDisplay: null,
-      esiJobTab: null,
-    },
-  },
-  {
-    groupID: null,
-    isSnapshot: false,
-    jobType: 1,
-    name: "Mining Drone I",
-    jobID: "job-1640798624845",
-    jobStatus: 4,
-    volume: 5,
-    itemID: 10246,
-    maxProductionLimit: 600,
-    runCount: 50,
-    jobCount: 1,
-    bpME: 0,
-    bpTE: 0,
-    structureType: 0,
-    rigType: 0,
-    systemType: 1,
-    apiJobs: new Set([477723892]),
-    apiOrders: new Set([6161413749]),
-    apiTransactions: new Set([
-      5225763813, 5763860401, 5763883247, 5763890268, 5763852213,
-    ]),
-    parentJob: [],
-    skills: [
-      {
-        level: 1,
-        typeID: 3380,
-      },
-    ],
-    rawData: {
-      materials: [
-        {
-          typeID: 40,
-          volume: 0.01,
-          jobType: 0,
-          quantity: 2,
-          name: "Megacyte",
-        },
-        {
-          volume: 0.01,
-          typeID: 34,
-          name: "Tritanium",
-          quantity: 77,
-          jobType: 0,
-        },
-        {
-          name: "Isogen",
-          quantity: 4,
-          volume: 0.01,
-          jobType: 0,
-          typeID: 37,
-        },
-        {
-          typeID: 38,
-          volume: 0.01,
-          name: "Nocxium",
-          quantity: 13,
-          jobType: 0,
-        },
-      ],
-      products: [
-        {
-          quantity: 1,
-          typeID: 10246,
-        },
-      ],
-      time: 300,
-    },
-    build: {
-      sale: {
-        brokersFee: [
-          {
-            order_id: 6161413749,
-            id: 20014198914,
-            complete: false,
-            date: "2021-12-30T18:33:45Z",
-            amount: 8550,
-            CharacterHash: "ABC123",
-          },
-        ],
-        marketOrders: [
-          {
-            duration: 90,
-            is_corporation: false,
-            issued: "2021-12-30T18:33:45Z",
-            location_id: 60008494,
-            location_name: "Amarr VIII (Oris) - Emperor Family Academy",
-            order_id: 6161413749,
-            item_price: 11400,
-            range: "region",
-            region_id: 10000043,
-            region_name: "Domain",
-            type_id: 10246,
-            item_name: "Mining Drone I",
-            volume_remain: 0,
-            volume_total: 50,
-            timeStamps: ["2021-12-30T18:33:45Z"],
-            CharacterHash: "ABC123",
-            complete: true,
-          },
-        ],
-        transactions: [
-          {
-            order_id: 6161413749,
-            journal_ref_id: 20016804153,
-            unit_price: 12000,
-            amount: 48000,
-            tax: 2150.4,
-            transaction_id: 5763852213,
-            quantity: 4,
-            date: "2021-12-31T14:40:06Z",
-            location_id: 60008494,
-            is_corp: false,
-            type_id: 10246,
-            description: "Abi Sweet",
-            CharacterHash: "ABC123",
-          },
-          {
-            order_id: 6161413749,
-            journal_ref_id: 20016841840,
-            unit_price: 11400,
-            amount: 57000,
-            tax: 2553.6,
-            transaction_id: 5763860401,
-            quantity: 5,
-            date: "2021-12-31T14:53:18Z",
-            location_id: 60008494,
-            is_corp: false,
-            type_id: 10246,
-            description: "IceQwen bought",
-            CharacterHash: "ABC123",
-          },
-          {
-            order_id: 6161413749,
-            journal_ref_id: 20016956492,
-            unit_price: 11400,
-            amount: 57000,
-            tax: 2553.6,
-            transaction_id: 5763883247,
-            quantity: 5,
-            date: "2021-12-31T15:35:22Z",
-            location_id: 60008494,
-            is_corp: false,
-            type_id: 10246,
-            description: "Igor Alexeevich",
-            CharacterHash: "ABC123",
-          },
-          {
-            order_id: 6161413749,
-            journal_ref_id: 20016991529,
-            unit_price: 11400,
-            amount: 228000,
-            tax: 10214.4,
-            transaction_id: 5763890268,
-            quantity: 20,
-            date: "2021-12-31T15:48:16Z",
-            location_id: 60008494,
-            is_corp: false,
-            type_id: 10246,
-            description: "Markus Dux",
-            CharacterHash: "ABC123",
-          },
-        ],
-        totalSold: 0,
-        totalSale: 0,
-        markUp: 0,
-      },
-      materials: [
-        {
-          name: "Megacyte",
-          typeID: 40,
-          quantityPurchased: 100,
-          purchaseComplete: true,
-          childJob: [],
-          purchasing: [
-            {
-              itemCount: 100,
-              itemCost: 769,
-              id: 1640798737317,
-            },
-          ],
-          quantity: 100,
-          volume: 0.01,
-          jobType: 0,
-          purchasedCost: 76900,
-        },
-        {
-          name: "Tritanium",
-          quantity: 3850,
-          jobType: 0,
-          childJob: [],
-          purchasing: [
-            {
-              id: 1640798745396,
-              itemCost: 3,
-              itemCount: 3850,
-            },
-          ],
-          purchasedCost: 11550,
-          volume: 0.01,
-          quantityPurchased: 3850,
-          purchaseComplete: true,
-          typeID: 34,
-        },
-        {
-          jobType: 0,
-          typeID: 37,
-          name: "Isogen",
-          childJob: [],
-          purchasing: [
-            {
-              itemCost: 136,
-              id: 1640798758521,
-              itemCount: 200,
-            },
-          ],
-          quantity: 200,
-          purchasedCost: 27200,
-          quantityPurchased: 200,
-          purchaseComplete: true,
-          volume: 0.01,
-        },
-        {
-          purchasing: [
-            {
-              itemCost: 1339,
-              id: 1640798770379,
-              itemCount: 650,
-            },
-          ],
-          childJob: [],
-          purchasedCost: 870350,
-          typeID: 38,
-          name: "Nocxium",
-          quantity: 650,
-          quantityPurchased: 650,
-          volume: 0.01,
-          purchaseComplete: true,
-          jobType: 0,
-        },
-      ],
-      products: {
-        quantityPerJob: 50,
-        recalculate: false,
-        totalQuantity: 50,
-      },
-      costs: {
-        installCosts: 92472,
-        totalPurchaseCost: 986000,
-        linkedJobs: [
-          {
-            status: "delivered",
-            CharacterHash: "ABC123",
-            runs: 50,
-            job_id: 477723892,
-            completed_date: "2021-12-29T21:29:31Z",
-            station_id: 60008494,
-            start_date: "2021-12-29T17:26:36Z",
-            end_date: "2021-12-29T20:46:36Z",
-            cost: 46236,
-            linked: false,
-            product_name: "Mining Drone I",
-            blueprint_type_id: 10247,
-            product_type_id: 10246,
-            activity_id: 1,
-            duration: 12000,
-          },
-        ],
-        extrasCosts: [],
-        extrasTotal: 0,
-      },
-      buildChar: "ABC123",
-      time: 10200,
-    },
-    layout: {
-      localMarketDisplay: null,
-      localOrderDisplay: null,
-      esiJobTab: null,
-    },
-  },
-  {
-    groupID: null,
-    isSnapshot: false,
-    jobType: 1,
-    name: "Light Missile Launcher I",
-    jobID: "job-1640798839177",
-    jobStatus: 4,
-    volume: 5,
-    itemID: 499,
-    maxProductionLimit: 200,
-    runCount: 20,
-    jobCount: 1,
-    bpME: 0,
-    bpTE: 0,
-    structureType: 0,
-    rigType: 0,
-    systemType: 1,
-    apiJobs: new Set([477724009]),
-    apiOrders: new Set(),
-    apiTransactions: new Set(),
-    parentJob: [],
-    skills: [
-      {
-        typeID: 3380,
-        level: 1,
-      },
-    ],
-    rawData: {
-      products: [
-        {
-          quantity: 1,
-          typeID: 499,
-        },
-      ],
-      materials: [
-        {
-          volume: 0.01,
-          name: "Tritanium",
-          jobType: 0,
-          quantity: 471,
-          typeID: 34,
-        },
-        {
-          jobType: 0,
-          quantity: 716,
-          typeID: 35,
-          name: "Pyerite",
-          volume: 0.01,
-        },
-      ],
-      time: 900,
-    },
-    build: {
-      products: {
-        quantityPerJob: 20,
-        totalQuantity: 20,
-        recalculate: false,
-      },
-      time: 12240,
-      materials: [
-        {
-          purchaseComplete: true,
-          typeID: 34,
-          name: "Tritanium",
-          quantityPurchased: 9420,
-          quantity: 9420,
-          volume: 0.01,
-          childJob: [],
-          purchasing: [
-            {
-              id: 1640798892235,
-              itemCost: 3.92,
-              itemCount: 9420,
-            },
-          ],
-          jobType: 0,
-          purchasedCost: 36926.4,
-        },
-        {
-          volume: 0.01,
-          jobType: 0,
-          quantityPurchased: 14320,
-          name: "Pyerite",
-          purchaseComplete: true,
-          typeID: 35,
-          childJob: [],
-          purchasing: [
-            {
-              itemCount: 14320,
-              itemCost: 13.37,
-              id: 1640798904818,
-            },
-          ],
-          purchasedCost: 191458.4,
-          quantity: 14320,
-        },
-      ],
-      sale: {
-        markUp: 0,
-        transactions: [
-          {
-            order_id: 6161414065,
-            journal_ref_id: 20012037654,
-            unit_price: 32950,
-            amount: 98850,
-            tax: 7908,
-            transaction_id: 5762870250,
-            quantity: 3,
-            date: "2021-12-30T01:42:27Z",
-            location_id: 60008494,
-            is_corp: false,
-            type_id: 499,
-            description: "ArianGG",
-            CharacterHash: "ABC123",
-          },
-          {
-            order_id: 6161414065,
-            journal_ref_id: 20011534029,
-            unit_price: 32950,
-            amount: 65900,
-            tax: 5272,
-            transaction_id: 5762760174,
-            quantity: 2,
-            date: "2021-12-29T21:53:26Z",
-            location_id: 60008494,
-            is_corp: false,
-            type_id: 499,
-            description: "Shelby Skye",
-            CharacterHash: "ABC123",
-          },
-          {
-            order_id: 6161414065,
-            journal_ref_id: 20018325638,
-            unit_price: 34250,
-            amount: 513750,
-            tax: 23016,
-            transaction_id: 5764179431,
-            quantity: 15,
-            date: "2022-01-01T02:36:32Z",
-            location_id: 60008494,
-            is_corp: false,
-            type_id: 499,
-            description: "Oryza Prodigy",
-            CharacterHash: "ABC123",
-          },
-        ],
-        totalSold: 0,
-        totalSale: 0,
-        brokersFee: [
-          {
-            order_id: 6161414065,
-            id: 20014197977,
-            complete: false,
-            date: "2021-12-30T18:33:26Z",
-            amount: 4860,
-            CharacterHash: "ABC123",
-          },
-        ],
-        marketOrders: [
-          {
-            duration: 90,
-            is_corporation: false,
-            issued: "2021-12-30T18:33:26Z",
-            location_id: 60008494,
-            location_name: "Amarr VIII (Oris) - Emperor Family Academy",
-            order_id: 6161414065,
-            item_price: 27000,
-            range: "region",
-            region_id: 10000043,
-            region_name: "Domain",
-            type_id: 499,
-            item_name: "Light Missile Launcher I",
-            volume_remain: 0,
-            volume_total: 20,
-            timeStamps: [],
-            CharacterHash: "ABC123",
-            complete: true,
-          },
-        ],
-      },
-      costs: {
-        installCosts: 35850,
-        extrasCosts: [],
-        extrasTotal: 0,
-        totalPurchaseCost: 228384.8,
-        linkedJobs: [
-          {
-            status: "delivered",
-            CharacterHash: "ABC123",
-            runs: 20,
-            job_id: 477724009,
-            completed_date: "2021-12-29T21:29:31Z",
-            station_id: 60008494,
-            start_date: "2021-12-29T17:28:32Z",
-            end_date: "2021-12-29T21:28:32Z",
-            cost: 17925,
-            linked: false,
-            product_name: "Light Missile Launcher I",
-            blueprint_type_id: 786,
-            product_type_id: 499,
-            activity_id: 1,
-            duration: 14400,
-          },
-        ],
-      },
-      buildChar: "ABC123",
-    },
-    layout: {
-      localMarketDisplay: null,
-      localOrderDisplay: null,
-      esiJobTab: null,
-    },
-  },
-  {
-    groupID: null,
-    jobType: 1,
-    name: "Reinforced Metallofullerene Alloys",
-    jobID: "job-1642710724535",
-    jobStatus: 0,
-    volume: 10,
-    itemID: 30008,
-    maxProductionLimit: 1000,
-    runCount: 50,
-    jobCount: 2,
-    bpME: 10,
-    bpTE: 10,
-    rigType: 2,
-    systemType: 2.1,
-    apiJobs: [],
-    apiOrders: [],
-    apiTransactions: [],
-    parentJob: [],
-    build: {
-      products: {
-        totalQuantity: 100,
-        quantityPerJob: 50,
-      },
-      costs: {
-        totalPurchaseCost: 0,
-        extrasCosts: [],
-        extrasTotal: 0,
-        linkedJobs: [],
-        installCosts: 0,
-      },
-      sale: {
-        totalSold: 0,
-        totalSale: 0,
-        markUp: 0,
-        marketOrders: [],
-        transactions: [],
-        brokersFee: [],
-      },
-      materials: [
-        {
-          jobType: 2,
-          typeID: 30305,
-          quantity: 940,
-          name: "Fullerene Intercalated Graphite",
-          volume: 0.8,
-          purchasing: [],
-          childJob: [],
-          quantityPurchased: 0,
-          purchasedCost: 0,
-          purchaseComplete: false,
-        },
-        {
-          jobType: 0,
-          typeID: 30019,
-          volume: 0.01,
-          name: "Powdered C-540 Graphite",
-          quantity: 172,
-          purchasing: [],
-          childJob: [],
-          quantityPurchased: 0,
-          purchasedCost: 0,
-          purchaseComplete: false,
-        },
-        {
-          jobType: 2,
-          name: "Scandium Metallofullerene",
-          typeID: 30308,
-          quantity: 1878,
-          volume: 0.65,
-          purchasing: [],
-          childJob: ["job-1642710994021"],
-          quantityPurchased: 0,
-          purchasedCost: 0,
-          purchaseComplete: false,
-        },
-        {
-          typeID: 30021,
-          name: "Modified Fluid Router",
-          quantity: 258,
-          volume: 0.01,
-          jobType: 0,
-          purchasing: [],
-          childJob: [],
-          quantityPurchased: 0,
-          purchasedCost: 0,
-          purchaseComplete: false,
-        },
-        {
-          typeID: 30022,
-          jobType: 0,
-          name: "Heuristic Selfassemblers",
-          quantity: 342,
-          volume: 0.01,
-          purchasing: [],
-          childJob: [],
-          quantityPurchased: 0,
-          purchasedCost: 0,
-          purchaseComplete: false,
-        },
-        {
-          jobType: 0,
-          name: "Electromechanical Hull Sheeting",
-          volume: 0.01,
-          quantity: 854,
-          typeID: 30254,
-          purchasing: [],
-          childJob: [],
-          quantityPurchased: 0,
-          purchasedCost: 0,
-          purchaseComplete: false,
-        },
-        {
-          typeID: 30309,
-          volume: 1.5,
-          name: "Graphene Nanoribbons",
-          jobType: 2,
-          quantity: 598,
-          purchasing: [],
-          childJob: ["job-1642711045845"],
-          quantityPurchased: 0,
-          purchasedCost: 0,
-          purchaseComplete: false,
-        },
-      ],
-      buildChar: "ABC123",
-      time: 11750,
-    },
-    rawData: {
-      materials: [
-        {
-          jobType: 2,
-          typeID: 30305,
-          quantity: 11,
-          name: "Fullerene Intercalated Graphite",
-          volume: 0.8,
-        },
-        {
-          jobType: 0,
-          typeID: 30019,
-          volume: 0.01,
-          name: "Powdered C-540 Graphite",
-          quantity: 2,
-        },
-        {
-          jobType: 2,
-          name: "Scandium Metallofullerene",
-          typeID: 30308,
-          quantity: 22,
-          volume: 0.65,
-        },
-        {
-          typeID: 30021,
-          name: "Modified Fluid Router",
-          quantity: 3,
-          volume: 0.01,
-          jobType: 0,
-        },
-        {
-          typeID: 30022,
-          jobType: 0,
-          name: "Heuristic Selfassemblers",
-          quantity: 4,
-          volume: 0.01,
-        },
-        {
-          jobType: 0,
-          name: "Electromechanical Hull Sheeting",
-          volume: 0.01,
-          quantity: 10,
-          typeID: 30254,
-        },
-        {
-          typeID: 30309,
-          volume: 1.5,
-          name: "Graphene Nanoribbons",
-          jobType: 2,
-          quantity: 7,
-        },
-      ],
-      products: [
-        {
-          typeID: 30008,
-          quantity: 1,
-        },
-      ],
-      time: 600,
-    },
-    structureType: 3,
-    skills: [
-      {
-        typeID: 3380,
-        level: 5,
-      },
-      {
-        typeID: 11441,
-        level: 2,
-      },
-      {
-        level: 2,
-        typeID: 11448,
-      },
-    ],
-    layout: {
-      localMarketDisplay: null,
-      localOrderDisplay: null,
-      esiJobTab: null,
-    },
-  },
-];
+export let jobArrayDefault = [];
 
-export let userJobSnapshotDefault = [
-  {
-    jobID: "job-1642030534709",
-    name: "Nova Light Missile",
-    runCount: 50,
-    jobCount: 1,
-    jobStatus: 2,
-    jobType: 1,
-    itemID: 213,
-    isSnapshot: false,
-    apiJobs: [],
-    apiOrders: [],
-    apiTransactions: [],
-    itemQuantity: 5000,
-    totalMaterials: 3,
-    totalComplete: 3,
-    buildVer: "0.6.0",
-    parentJob: [],
-    childJobs: [],
-    materialIDs: [34, 35, 36],
-    metaLevel: null,
-    projectID: null,
-    endDateDisplay: null,
-  },
-  {
-    jobID: "job-1642710994021",
-    name: "Scandium Metallofullerene",
-    runCount: 12,
-    jobCount: 1,
-    jobStatus: 1,
-    jobType: 2,
-    itemID: 30308,
-    isSnapshot: false,
-    apiJobs: [],
-    apiOrders: [],
-    apiTransactions: [],
-    itemQuantity: 1920,
-    totalMaterials: 4,
-    totalComplete: 0,
-    buildVer: "0.6.0",
-    parentJob: [1642710724535],
-    childJobs: [],
-    materialIDs: [4247, 39, 30373, 30375],
-    metaLevel: null,
-    projectID: null,
-    endDataDisplay: null,
-  },
-  {
-    jobID: "job-1642711045845",
-    name: "Graphene Nanoribbons",
-    runCount: 5,
-    jobCount: 1,
-    jobStatus: 1,
-    jobType: 2,
-    itemID: 30309,
-    isSnapshot: false,
-    apiJobs: [],
-    apiOrders: [],
-    apiTransactions: [],
-    itemQuantity: 600,
-    totalMaterials: 4,
-    totalComplete: 0,
-    buildVer: "0.6.0",
-    parentJob: [1642710724535],
-    childJobs: [],
-    materialIDs: [4051, 30375, 30376, 38],
-    metaLevel: null,
-    projectID: null,
-    endDataDisplay: null,
-  },
-  {
-    jobID: "job-1640798624845",
-    name: "Mining Drone I",
-    runCount: 50,
-    jobCount: 1,
-    jobStatus: 4,
-    jobType: 1,
-    itemID: 10246,
-    isSnapshot: false,
-    apiJobs: [477723892],
-    apiOrders: [6161413749],
-    apiTransactions: [5225763813, 5763860401, 5763883247, 5763890268],
-    itemQuantity: 50,
-    totalMaterials: 4,
-    totalComplete: 4,
-    buildVer: "0.6.0",
-    parentJob: [],
-    childJobs: [],
-    materialIDs: [40, 34, 37, 38],
-    metaLevel: null,
-    projectID: null,
-    endDateDisplay: "2021-12-29T20:46:36Z",
-  },
-  {
-    jobID: "job-1642710724535",
-    name: "Reinforced Metallofullerene Alloys",
-    runCount: 50,
-    jobCount: 2,
-    jobStatus: 0,
-    jobType: 1,
-    itemID: 30008,
-    isSnapshot: false,
-    apiJobs: [],
-    apiOrders: [],
-    apiTransactions: [],
-    itemQuantity: 100,
-    totalMaterials: 7,
-    totalComplete: 0,
-    buildVer: "0.6.0",
-    parentJob: [],
-    childJobs: [1642710994021, 1642711045845],
-    materialIDs: [30305, 30019, 30308, 30021, 30022, 30309, 30254],
-    projectID: null,
-    endDateDisplay: null,
-  },
-];
+export let userJobSnapshotDefault = [];
 
 export let jobStatusDefault = [
   {
@@ -1288,399 +48,326 @@ export let jobStatusDefault = [
 export let apiJobsDefault = [
   {
     user: "ABC123",
-    data: [
-      {
-        activity_id: 1,
-        blueprint_id: 1037980236955,
-        blueprint_location_id: 60008494,
-        blueprint_type_id: 10247,
-        completed_character_id: 94800326,
-        completed_date: "2021-12-29T21:29:31Z",
-        cost: 46236,
-        duration: 12000,
-        end_date: "2021-12-29T20:46:36Z",
-        facility_id: 60008494,
-        installer_id: 94800326,
-        job_id: 477723892,
-        licensed_runs: 600,
-        output_location_id: 60008494,
-        probability: 1,
-        product_type_id: 10246,
-        runs: 50,
-        start_date: "2021-12-29T17:26:36Z",
-        station_id: 60008494,
-        status: "delivered",
-        successful_runs: 50,
-        product_name: "Mining Drone I",
-        linked: true,
-      },
-      {
-        activity_id: 1,
-        blueprint_id: 1020888478080,
-        blueprint_location_id: 60008494,
-        blueprint_type_id: 786,
-        completed_character_id: 94800326,
-        completed_date: "2021-12-29T21:29:31Z",
-        cost: 17925,
-        duration: 14400,
-        end_date: "2021-12-29T21:28:32Z",
-        facility_id: 60008494,
-        installer_id: 94800326,
-        job_id: 477724009,
-        licensed_runs: 200,
-        output_location_id: 60008494,
-        probability: 1,
-        product_type_id: 499,
-        runs: 20,
-        start_date: "2021-12-29T17:28:32Z",
-        station_id: 60008494,
-        status: "delivered",
-        successful_runs: 20,
-        product_name: "Light Missile Launcher I",
-        linked: true,
-      },
-      {
-        activity_id: 1,
-        blueprint_id: 1019652411598,
-        blueprint_location_id: 60008494,
-        blueprint_type_id: 814,
-        cost: 3064,
-        duration: 12000,
-        end_date: "2021-01-08T03:17:02Z",
-        facility_id: 60008494,
-        installer_id: 94800326,
-        job_id: 476337809,
-        licensed_runs: 600,
-        output_location_id: 60008494,
-        probability: 1,
-        product_type_id: 213,
-        runs: 50,
-        start_date: "2021-01-06T23:57:02Z",
-        station_id: 60008494,
-        status: "active",
-        product_name: "Nova Light Missile",
-        linked: false,
-      },
-    ],
+    data: [],
   },
 ];
 
 export let defaultEsiSkills = [
   {
     user: "ABC123",
-    data: [
-      {
+    data: {
+      3395: {
         id: 3395,
         name: "Advanced Small Ship Construction",
         activeLevel: 5,
       },
-      {
+      24625: {
         id: 24625,
         name: "Advanced Mass Production",
         activeLevel: 5,
       },
-      {
+      45749: {
         id: 45749,
         name: "Advanced Mass Reactions",
         activeLevel: 5,
       },
-      {
+      3397: {
         id: 3397,
         name: "Advanced Medium Ship Construction",
         activeLevel: 5,
       },
-      {
+      24624: {
         id: 24624,
         name: "Advanced Laboratory Operation",
         activeLevel: 5,
       },
-      {
+      3398: {
         id: 3398,
         name: "Advanced Large Ship Construction",
         activeLevel: 5,
       },
-      {
+      3396: {
         id: 3396,
         name: "Advanced Industrial Ship Construction",
         activeLevel: 5,
       },
-      {
+      3388: {
         id: 3388,
         name: "Advanced Industry",
         activeLevel: 5,
       },
-      {
+      23087: {
         id: 23087,
         name: "Amarr Encryption Methods",
         activeLevel: 3,
       },
-      {
+      11444: {
         id: 11444,
         name: "Amarr Starship Engineering",
         activeLevel: 5,
       },
-      {
+      26253: {
         id: 26253,
         name: "Armor Rigging",
         activeLevel: 5,
       },
-      {
+      26254: {
         id: 26254,
         name: "Astronautics Rigging",
         activeLevel: 5,
       },
-      { name: "Broker Relations", id: 3446, activeLevel: 5 },
-      {
+      3446: { name: "Broker Relations", id: 3446, activeLevel: 5 },
+      11454: {
         id: 11454,
         name: "Caldari Starship Engineering",
         activeLevel: 5,
       },
-      {
+      21790: {
         id: 21790,
         name: "Caldari Encryption Methods",
         activeLevel: 5,
       },
-      {
+      22242: {
         id: 22242,
         name: "Capital Ship Consruction",
         activeLevel: 5,
       },
-      {
+      30325: {
         id: 30325,
         name: "Core Subsystem Technology",
         activeLevel: 5,
       },
-      {
+      30324: {
         id: 30324,
         name: "Defensive Subsystem Technology",
         activeLevel: 5,
       },
-      {
+      26255: {
         id: 26255,
         name: "Drones Rigging",
         activeLevel: 5,
       },
-      {
+      26224: {
         id: 26224,
         name: "Drug Manufacturing",
         activeLevel: 5,
       },
-      {
+      11448: {
         id: 11448,
         name: "Electromagnetic Physics",
         activeLevel: 5,
       },
-      {
+      11453: {
         id: 11453,
         name: "Electronic Engineering",
         activeLevel: 5,
       },
-      {
+      26256: {
         id: 26256,
         name: "Electronic Superiority Rigging",
         activeLevel: 5,
       },
-      {
+      26258: {
         id: 26258,
         name: "Energy Weapon Rigging",
         activeLevel: 5,
       },
-      {
+      23121: {
         id: 23121,
         name: "Gallente Encryption Methods",
         activeLevel: 5,
       },
-      {
+      11450: {
         id: 11450,
         name: "Gallente Starship Engineering",
         activeLevel: 5,
       },
-      {
+      11446: {
         id: 11446,
         name: "Graviton Physics",
         activeLevel: 5,
       },
-      {
+      11433: {
         id: 11433,
         name: "High Energy Physics",
         activeLevel: 5,
       },
-      {
+      26259: {
         id: 26259,
         name: "Hybrid Weapon Rigging",
         activeLevel: 5,
       },
-      {
+      11443: {
         id: 11443,
         name: "Hydromagnetic Physics",
         activeLevel: 5,
       },
-      {
+      3380: {
         id: 3380,
         name: "Industry",
         activeLevel: 5,
       },
-      {
+      26252: {
         id: 26252,
         name: "Jury Rigging",
         activeLevel: 5,
       },
-      {
+      3406: {
         id: 3406,
         name: "Laboratory Operation",
         activeLevel: 5,
       },
-      {
+      11447: {
         id: 11447,
         name: "Laser Physics",
         activeLevel: 5,
       },
-      {
+      26260: {
         id: 26260,
         name: "Launcher Rigging",
         activeLevel: 5,
       },
-      {
+      3387: {
         id: 3387,
         name: "Mass Production",
         activeLevel: 5,
       },
-      {
+      45748: {
         id: 45748,
         name: "Mass Reactions",
         activeLevel: 5,
       },
-      {
+      11452: {
         id: 11452,
         name: "Mechanical Engineering",
         activeLevel: 5,
       },
-      {
+      3409: {
         id: 3409,
         name: "Metallurgy",
         activeLevel: 5,
       },
-      {
+      21791: {
         id: 21791,
         name: "Minmatar Encryption Methods",
         activeLevel: 5,
       },
-      {
+      11445: {
         id: 11445,
         name: "Minmatar Starship Engineering",
         activeLevel: 5,
       },
-      {
+      11529: {
         id: 11529,
         name: "Molecular Engineering",
         activeLevel: 5,
       },
-      {
+      11442: {
         id: 11442,
         name: "Nanite Engineering",
         activeLevel: 5,
       },
-      {
+      11451: {
         id: 11451,
         name: "Nuclear Physics",
         activeLevel: 5,
       },
-      {
+      30327: {
         id: 30327,
         name: "Offensive Subsystem Technology",
         activeLevel: 5,
       },
-      {
+      3400: {
         id: 3400,
         name: "Outpost Construction",
         activeLevel: 5,
       },
-      {
+      26257: {
         id: 26257,
         name: "Projectile Weapon Rigging",
         activeLevel: 5,
       },
-      {
+      30788: {
         id: 30788,
         name: "Propulsion Subsystem Technology",
         activeLevel: 5,
       },
-      {
+      11441: {
         id: 11441,
         name: "Plasma Physics",
         activeLevel: 5,
       },
-      {
+      11455: {
         id: 11455,
         name: "Quantum Physics",
         activeLevel: 5,
       },
-      {
+      45746: {
         id: 45746,
         name: "Reactions",
         activeLevel: 5,
       },
-      {
+      3403: {
         id: 3403,
         name: "Research",
         activeLevel: 5,
       },
-      {
+      11449: {
         id: 11449,
         name: "Rocket Science",
         activeLevel: 5,
       },
-      {
+      3402: {
         id: 3402,
         name: "Science",
         activeLevel: 5,
       },
-      {
+      26261: {
         id: 26261,
         name: "Shield Rigging",
         activeLevel: 5,
       },
-      {
+      3408: {
         id: 3408,
         name: "Sleeper Encryption Methods",
         activeLevel: 5,
       },
-      {
+      21789: {
         id: 21789,
         name: "Sleeper Technology",
         activeLevel: 5,
       },
-      {
+      20433: {
         id: 20433,
         name: "Talocan Technology",
         activeLevel: 5,
       },
-      {
+      23123: {
         id: 23123,
         name: "Takmahl Technology",
         activeLevel: 5,
       },
-      {
+      52308: {
         id: 52308,
         name: "Triglavian Encryption Methods",
         activeLevel: 5,
       },
-      {
+      52307: {
         id: 52307,
         name: "Triglavian Quantum Engineering",
         activeLevel: 5,
       },
-      {
+      55025: {
         id: 55025,
         name: "Upwell Encryption Methods",
         activeLevel: 5,
       },
-      {
+      23124: {
         id: 23124,
         name: "Yan Jung Technology",
         activeLevel: 5,
       },
-    ],
+    },
   },
 ];
 
@@ -1701,44 +388,7 @@ export let defaultEsiOrders = [
 export let defaultEsiHistOrders = [
   {
     user: "ABC123",
-    data: [
-      {
-        duration: 90,
-        is_corporation: false,
-        issued: "2021-12-30T18:33:45Z",
-        location_id: 60008494,
-        order_id: 6161413749,
-        price: 11400,
-        range: "region",
-        region_id: 10000043,
-        state: "expired",
-        type_id: 10246,
-        volume_remain: 0,
-        volume_total: 50,
-        user_id: 94800326,
-        location_name: "Amarr VIII (Oris) - Emperor Family Academy",
-        region_name: "Domain",
-        item_name: "Mining Drone I",
-      },
-      {
-        duration: 90,
-        is_corporation: false,
-        issued: "2021-12-30T18:33:26Z",
-        location_id: 60008494,
-        order_id: 6161414065,
-        price: 27000,
-        range: "region",
-        region_id: 10000043,
-        state: "expired",
-        type_id: 499,
-        volume_remain: 0,
-        volume_total: 20,
-        user_id: 94800326,
-        location_name: "Amarr VIII (Oris) - Emperor Family Academy",
-        region_name: "Domain",
-        item_name: "Light Missile Launcher I",
-      },
-    ],
+    data: [],
   },
 ];
 
@@ -1752,395 +402,14 @@ export let defaultEsiBlueprints = [
 export let defaultEsiTransactions = [
   {
     user: "ABC123",
-    data: [
-      {
-        CharacterHash: "ABC123",
-        client_id: 784956701,
-        date: "2021-12-31T15:48:16Z",
-        is_buy: false,
-        is_personal: true,
-        journal_ref_id: 20016991529,
-        location_id: 60008494,
-        quantity: 20,
-        transaction_id: 5763890268,
-        type_id: 10246,
-        unit_price: 11400,
-      },
-      {
-        CharacterHash: "ABC123",
-        client_id: 2119194221,
-        date: "2021-12-31T15:35:22Z",
-        is_buy: false,
-        is_personal: true,
-        journal_ref_id: 20016956492,
-        location_id: 60008494,
-        quantity: 5,
-        transaction_id: 5763883247,
-        type_id: 10246,
-        unit_price: 11400,
-      },
-      {
-        CharacterHash: "ABC123",
-        client_id: 2118912977,
-        date: "2021-12-31T14:53:18Z",
-        is_buy: false,
-        is_personal: true,
-        journal_ref_id: 20016841840,
-        location_id: 60008494,
-        quantity: 5,
-        transaction_id: 5763860401,
-        type_id: 10246,
-        unit_price: 11400,
-      },
-      {
-        CharacterHash: "ABC123",
-        client_id: 2119059949,
-        date: "2021-12-31T14:40:06Z",
-        is_buy: false,
-        is_personal: true,
-        journal_ref_id: 20016804153,
-        location_id: 60008494,
-        quantity: 4,
-        transaction_id: 5763852213,
-        type_id: 10246,
-        unit_price: 12000,
-      },
-      {
-        CharacterHash: "ABC123",
-        client_id: 96311212,
-        date: "2021-12-30T18:39:02Z",
-        is_buy: false,
-        is_personal: true,
-        journal_ref_id: 20014214855,
-        location_id: 60008494,
-        quantity: 5,
-        transaction_id: 5763314805,
-        type_id: 10246,
-        unit_price: 11400,
-      },
-      {
-        CharacterHash: "ABC123",
-        client_id: 2119433506,
-        date: "2021-12-29T23:00:46Z",
-        is_buy: false,
-        is_personal: true,
-        journal_ref_id: 20011710120,
-        location_id: 60008494,
-        quantity: 10,
-        transaction_id: 5762799195,
-        type_id: 10246,
-        unit_price: 11850,
-      },
-      {
-        CharacterHash: "ABC123",
-        client_id: 122887441,
-        date: "2022-01-01T02:36:32Z",
-        is_buy: false,
-        is_personal: true,
-        journal_ref_id: 20018325638,
-        location_id: 60008494,
-        quantity: 15,
-        transaction_id: 5764179431,
-        type_id: 499,
-        unit_price: 34250,
-      },
-      {
-        CharacterHash: "ABC123",
-        client_id: 2119513212,
-        date: "2021-12-30T01:42:27Z",
-        is_buy: false,
-        is_personal: true,
-        journal_ref_id: 20012037654,
-        location_id: 60008494,
-        quantity: 3,
-        transaction_id: 5762870250,
-        type_id: 499,
-        unit_price: 32950,
-      },
-      {
-        CharacterHash: "ABC123",
-        client_id: 2119484455,
-        date: "2021-12-29T21:53:26Z",
-        is_buy: false,
-        is_personal: true,
-        journal_ref_id: 20011534029,
-        location_id: 60008494,
-        quantity: 2,
-        transaction_id: 5762760174,
-        type_id: 499,
-        unit_price: 32950,
-      },
-    ],
+    data: [],
   },
 ];
 
 export let defaultEsiJournal = [
   {
     user: "ABC123",
-    data: [
-      {
-        CharacterHash: "ABC123",
-        amount: -5472,
-        balance: 5835304896.8761,
-        date: "2021-12-30T18:33:45Z",
-        description:
-          "Market order commission to broker authorized by: Reginal Shardani",
-        first_party_id: 94800326,
-        id: 20014198914,
-        reason: "",
-        ref_type: "brokers_fee",
-        second_party_id: 1000086,
-      },
-      {
-        CharacterHash: "ABC123",
-        amount: -4860,
-        balance: 5835310368.8761,
-        date: "2021-12-30T18:33:26Z",
-        description:
-          "Market order commission to broker authorized by: Reginal Shardani",
-        first_party_id: 94800326,
-        id: 20014197977,
-        reason: "",
-        ref_type: "brokers_fee",
-        second_party_id: 1000086,
-      },
-      {
-        CharacterHash: "ABC123",
-        amount: 65900,
-        balance: 5835120538.8761,
-        context_id: 5762760174,
-        context_id_type: "market_transaction_id",
-        date: "2021-12-29T21:53:26Z",
-        description:
-          "Market: Shelby Skye Singularity bought stuff from Reginal Shardani",
-        first_party_id: 2119484455,
-        id: 20011534027,
-        reason: "",
-        ref_type: "market_transaction",
-        second_party_id: 94800326,
-      },
-      {
-        CharacterHash: "ABC123",
-        amount: -5272,
-        balance: 5835115266.8761,
-        date: "2021-12-29T21:53:26Z",
-        description: "Sales tax paid to the SCC",
-        first_party_id: 94800326,
-        id: 20011534028,
-        reason: "",
-        ref_type: "transaction_tax",
-        second_party_id: 1000132,
-      },
-      {
-        CharacterHash: "ABC123",
-        amount: 98850,
-        balance: 5835323136.8761,
-        context_id: 5762870250,
-        context_id_type: "market_transaction_id",
-        date: "2021-12-30T01:42:27Z",
-        description: "Market: ArianGG bought stuff from Reginal Shardani",
-        first_party_id: 2119513212,
-        id: 20012037652,
-        reason: "",
-        ref_type: "market_transaction",
-        second_party_id: 94800326,
-      },
-      {
-        CharacterHash: "ABC123",
-        amount: -7908,
-        balance: 5835315228.8761,
-        date: "2021-12-30T01:42:27Z",
-        description: "Sales tax paid to the SCC",
-        first_party_id: 94800326,
-        id: 20012037653,
-        reason: "",
-        ref_type: "transaction_tax",
-        second_party_id: 1000132,
-      },
-      {
-        CharacterHash: "ABC123",
-        amount: 513750,
-        balance: 4166935940.1241,
-        context_id: 5764179431,
-        context_id_type: "market_transaction_id",
-        date: "2022-01-01T02:36:32Z",
-        description: "Market: Oryza Prodigy bought stuff from Reginal Shardani",
-        first_party_id: 122887441,
-        id: 20018325636,
-        reason: "",
-        ref_type: "market_transaction",
-        second_party_id: 94800326,
-      },
-      {
-        CharacterHash: "ABC123",
-        amount: -23016,
-        balance: 4166912924.1241,
-        date: "2022-01-01T02:36:32Z",
-        description: "Sales tax paid to the SCC",
-        first_party_id: 94800326,
-        id: 20018325637,
-        reason: "",
-        ref_type: "transaction_tax",
-        second_party_id: 1000132,
-      },
-      {
-        CharacterHash: "ABC123",
-        amount: 118500,
-        balance: 5835233766.8761,
-        context_id: 5762799195,
-        context_id_type: "market_transaction_id",
-        date: "2021-12-29T23:00:46Z",
-        description: "Market: Tex Ruxim bought stuff from Reginal Shardani",
-        first_party_id: 2119433506,
-        id: 20011710116,
-        reason: "",
-        ref_type: "market_transaction",
-        second_party_id: 94800326,
-      },
-      {
-        CharacterHash: "ABC123",
-        amount: -9480,
-        balance: 5835224286.8761,
-        date: "2021-12-29T23:00:46Z",
-        description: "Sales tax paid to the SCC",
-        first_party_id: 94800326,
-        id: 20011710118,
-        reason: "",
-        ref_type: "transaction_tax",
-        second_party_id: 1000132,
-      },
-      {
-        CharacterHash: "ABC123",
-        amount: 57000,
-        balance: 5835361896.8761,
-        context_id: 5763314805,
-        context_id_type: "market_transaction_id",
-        date: "2021-12-30T18:39:02Z",
-        description: "Market: Griff Blad bought stuff from Reginal Shardani",
-        first_party_id: 96311212,
-        id: 20014214853,
-        reason: "",
-        ref_type: "market_transaction",
-        second_party_id: 94800326,
-      },
-      {
-        CharacterHash: "ABC123",
-        amount: -4560,
-        balance: 5835357336.8761,
-        date: "2021-12-30T18:39:02Z",
-        description: "Sales tax paid to the SCC",
-        first_party_id: 94800326,
-        id: 20014214854,
-        reason: "",
-        ref_type: "transaction_tax",
-        second_party_id: 1000132,
-      },
-      {
-        CharacterHash: "ABC123",
-        amount: 48000,
-        balance: 4166086686.8761,
-        context_id: 5763852213,
-        context_id_type: "market_transaction_id",
-        date: "2021-12-31T14:40:06Z",
-        description: "Market: Abi Sweet bought stuff from Reginal Shardani",
-        first_party_id: 2119059949,
-        id: 20016804151,
-        reason: "",
-        ref_type: "market_transaction",
-        second_party_id: 94800326,
-      },
-      {
-        CharacterHash: "ABC123",
-        amount: -2150.4,
-        balance: 4166084536.4761,
-        date: "2021-12-31T14:40:06Z",
-        description: "Sales tax paid to the SCC",
-        first_party_id: 94800326,
-        id: 20016804152,
-        reason: "",
-        ref_type: "transaction_tax",
-        second_party_id: 1000132,
-      },
-      {
-        CharacterHash: "ABC123",
-        amount: 57000,
-        balance: 4166141536.4761,
-        context_id: 5763860401,
-        context_id_type: "market_transaction_id",
-        date: "2021-12-31T14:53:18Z",
-        description: "Market: IceQwen bought stuff from Reginal Shardani",
-        first_party_id: 2118912977,
-        id: 20016841838,
-        reason: "",
-        ref_type: "market_transaction",
-        second_party_id: 94800326,
-      },
-      {
-        CharacterHash: "ABC123",
-        amount: -2553.6,
-        balance: 4166138982.8761,
-        date: "2021-12-31T14:53:18Z",
-        description: "Sales tax paid to the SCC",
-        first_party_id: 94800326,
-        id: 20016841839,
-        reason: "",
-        ref_type: "transaction_tax",
-        second_party_id: 1000132,
-      },
-      {
-        CharacterHash: "ABC123",
-        amount: 57000,
-        balance: 4166195982.8761,
-        context_id: 5763883247,
-        context_id_type: "market_transaction_id",
-        date: "2021-12-31T15:35:22Z",
-        description:
-          "Market: Igor Alexeevich bought stuff from Reginal Shardani",
-        first_party_id: 2119194221,
-        id: 20016956490,
-        reason: "",
-        ref_type: "market_transaction",
-        second_party_id: 94800326,
-      },
-      {
-        CharacterHash: "ABC123",
-        amount: -2553.6,
-        balance: 4166193429.2761,
-        date: "2021-12-31T15:35:22Z",
-        description: "Sales tax paid to the SCC",
-        first_party_id: 94800326,
-        id: 20016956491,
-        reason: "",
-        ref_type: "transaction_tax",
-        second_party_id: 1000132,
-      },
-      {
-        CharacterHash: "ABC123",
-        amount: 228000,
-        balance: 4166421429.2761,
-        context_id: 5763890268,
-        context_id_type: "market_transaction_id",
-        date: "2021-12-31T15:48:16Z",
-        description: "Market: Markus Dux bought stuff from Reginal Shardani",
-        first_party_id: 784956701,
-        id: 20016991527,
-        reason: "",
-        ref_type: "market_transaction",
-        second_party_id: 94800326,
-      },
-      {
-        CharacterHash: "ABC123",
-        amount: -10214.4,
-        balance: 4166411214.8761,
-        date: "2021-12-31T15:48:16Z",
-        description: "Sales tax paid to the SCC",
-        first_party_id: 94800326,
-        id: 20016991528,
-        reason: "",
-        ref_type: "transaction_tax",
-        second_party_id: 1000132,
-      },
-    ],
+    data: [],
   },
 ];
 export let defaultEsiAssets = [];
@@ -2175,25 +444,15 @@ export let usersDefault = [
         hideCompleteMaterials: false,
         citadelBrokersFee: 1,
       },
+      structures: {
+        manufacturing: [],
+        reaction: [],
+      },
     },
   },
 ];
 
 export let apiOrdersDefault = [
-  {
-    duration: 90,
-    is_corporation: false,
-    issued: "2021-12-30T18:33:45Z",
-    location_id: 60008494,
-    order_id: 6161413749,
-    price: 11400,
-    range: "region",
-    region_id: 10000043,
-    state: "expired",
-    type_id: 10246,
-    volume_remain: 0,
-    volume_total: 50,
-  },
 ];
 
 export let eveIDsDefault = [
@@ -2251,7 +510,7 @@ export let blueprintOptions = {
 };
 export let structureOptions = {
   manStructure: {
-    0: { id: 0, label: "NPC Station", material: 0, time: 0, cost: 0 },
+    0: { id: 0, label: "NPC Station", material: 0, time: 0, cost: 0, defaultTax: 0.25 },
     1: { id: 1, label: "Medium", material: 1, time: 0.15, cost: 0.03 },
     2: { id: 2, label: "Large", material: 1, time: 0.2, cost: 0.04 },
     3: { id: 3, label: "X-Large", material: 1, time: 0.3, cost: 0.05 },
