@@ -7,20 +7,16 @@ import {
 } from "react-router-dom";
 import { Header } from "./Components/Header";
 import { Home } from "./Components/Landing Page";
-import { ItemTree } from "./Components/item Tree";
 import { Footer } from "./Components/Footer/Footer";
 import { lazy, Suspense, useContext } from "react";
 import { JobPlanner } from "./Components/Job Planner";
 import { IsLoggedInContext } from "./Context/AuthContext";
-import { Container } from "@mui/material";
 import { LoadingPage } from "./Components/loadingPage";
-import EditJob from "./Components/Job Planner/Edit Job/EditJob";
-import { EditJob_New } from "./Components/Edit Job/editJob";
+
+// import { EditJob_New } from "./Components/Edit Job/editJob";
 
 const AuthMainUser = lazy(() => import("./Components/Auth/MainUserAuth"));
-const UpcomingChanges = lazy(() =>
-  import("./Components/Upcoming Changes/upcomingReleases")
-);
+const EditJobPage = lazy(() => import("./Components/Edit Job/editJob"));
 const AccountsPage = lazy(() => import("./Components/Accounts/Accounts"));
 const SettingsPage = lazy(() => import("./Components/Settings/Settings"));
 const BlueprintLibrary = lazy(() =>
@@ -32,32 +28,24 @@ const Dashboard = lazy(() => import("./Components/Dashboard/Dashboard"));
 export function NavRoutes({ mode, colorMode }) {
   return (
     <BrowserRouter>
-      <Container
-        disableGutters
-        maxWidth="false"
-        sx={{ marigin: "0px", padding: "0px" }}
-      >
-        <Header mode={mode} colorMode={colorMode} />
-        <Suspense fallback={<LoadingPage />}>
-          <Routes>
-            <Route path="/" element={<Home />} />
-            <Route path="/jobplanner" element={<JobPlanner />} />
-            <Route path="/editjob/:jobID" element={<EditJob_New />} />
-            <Route path="/auth/" element={<AuthMainUser />} />
-            {/* <Route path="/itemtree" element={<ItemTree />} /> */}
-            <Route path="/upcoming-changes" element={<UpcomingChanges />} />
-            <Route element={<ProtectedRoute />}>
-              <Route path="/dashboard" element={<Dashboard />} />
-              <Route path="/accounts" element={<AccountsPage />} />
-              <Route path="/settings" element={<SettingsPage />} />
-              <Route path="/blueprint-library" element={<BlueprintLibrary />} />
-              <Route path="/asset-library" element={<AssetLibrary />} />
-            </Route>
-            <Route path="*" element={<Navigate to="/" />} />
-          </Routes>
-        </Suspense>
-        <Footer />
-      </Container>
+      <Header mode={mode} colorMode={colorMode} />
+      <Suspense fallback={<LoadingPage />}>
+        <Routes>
+          <Route path="/" element={<Home />} />
+          <Route path="/jobplanner" element={<JobPlanner />} />
+          <Route path="/editjob/:jobID" element={<EditJobPage />} />
+          <Route path="/auth/" element={<AuthMainUser />} />
+          <Route element={<ProtectedRoute />}>
+            <Route path="/dashboard" element={<Dashboard />} />
+            <Route path="/accounts" element={<AccountsPage />} />
+            <Route path="/settings" element={<SettingsPage />} />
+            <Route path="/blueprint-library" element={<BlueprintLibrary />} />
+            <Route path="/asset-library" element={<AssetLibrary />} />
+          </Route>
+          <Route path="*" element={<Navigate to="/" />} />
+        </Routes>
+      </Suspense>
+      <Footer />
     </BrowserRouter>
   );
 }
